+++ src/main/java/com/study/domain/Lottos.java
@@ -1,4 +1,4 @@
-package domain;
+package com.study.domain;
 
 import java.util.List;
